take the entity string tests out of tests/risks
